#Issue 293
Incremented CharacterData.class_version in chardata_definitions.h so that cereal throws a critical error instead of crashing the authserver

Issue #352

Removed unused codes in MapInstance.cpp and changed size_t into uint32_t because size_t works differently between 32bit and 64bit systems

+++ Projects/CoX/Common/GameData/DataStorage.cpp
@@ -74,7 +74,7 @@ bool BinStore::read_data_blocks( bool file_data_blocks )
         return true;
     }
     QString hdr=read_pstr(20);
-    size_t sz;
+    uint32_t sz;
     read_internal(sz);
 
     quint64 read_start = m_str.pos();
